Synchronize with OpenSC SVN r5337
Warn c89 fix on cwa1890.c



git-svn-id: https://forja.cenatic.es/svn/opendnie@293 52e26522-a1c9-48c4-a7af-28ea70901177

+++ opensc-opendnie/trunk/src/libopensc/cwa14890.c
@@ -1358,7 +1358,7 @@ int cwa_encode_apdu(sc_card_t * card,
 	DES_key_schedule k2;
 	char *msg = NULL;
 
-	int i, j;		/* for xor loops */
+	size_t i, j;		/* for xor loops */
 	int res = SC_SUCCESS;
 	sc_context_t *ctx = NULL;
 	cwa_sm_status_t *sm = NULL;
